Move character select into dim-ui
