+++ src/app/d2-loadout-builder/LoadoutBuilder.tsx
@@ -5,7 +5,7 @@ import _ from 'lodash';
 import React from 'react';
 import { connect } from 'react-redux';
 import { DestinyAccount } from '../accounts/destiny-account.service';
-import CharacterSelect from '../character-select/CharacterSelect';
+import CharacterSelect from '../dim-ui/CharacterSelect';
 import { Loading } from '../dim-ui/Loading';
 import { D2StoresService } from '../inventory/d2-stores.service';
 import { DimStore, D2Store } from '../inventory/store-types';
